collector/pyproject.toml
[project]
name = "bowser-collector"
version = "0.1.0"
description = "Bowser v1 data collector — pulls Yahoo Finance, Gaspy, MBIE, news, FX into snapshot.json"
requires-python = ">=3.11"
dependencies = [
    "httpx>=0.27",
    "feedparser>=6.0.11",
    "pydantic>=2.7",
    "pandas>=2.2",
    "python-dotenv>=1.0",
    "apscheduler>=3.10",
    "google-cloud-storage>=2.18",
]

[dependency-groups]
dev = [
    "pytest>=8.2",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-q"
